refactor(rimap-content): extract html/pipeline.rs from html/mod.rs
Moves HtmlResult, HiddenMethod (+ impl), the 5 pub(crate) constants
(MAX_HTML_BYTES, MAX_ANCHOR_TEXT_SCAN, MAX_HIDDEN_HITS,
MAX_MISMATCH_HITS, HTML_BODY_LIMIT_KIND), the ElementIndex type
alias, the sanitize orchestrator, and the tests module from
html/mod.rs (914 LOC) to a new html/pipeline.rs sibling (903 LOC).
html/mod.rs shrinks to 25 LOC of pure wiring + re-exports.

mod.rs re-exports the public types (HtmlResult + sanitize) plus the
pub(crate) items siblings consume via `crate::html::*` (extract.rs,
hidden.rs, mismatch.rs, style_parse.rs all expect this path). No
public-API change.

No behavior change. All 1416 workspace tests still pass.

Refs #300 (item 3, hub 4/4)
